chore: 패키지 내 불필요 명칭 전부 제거 step관련 명칭 패키지내에서 제거
