+++ src/main/java/racingcarcompetition/Main.java
@@ -1,0 +1,8 @@
+package racingcarcompetition;
+
+public class Main {
+    public static void main(String[] args) {
+        RacingCarCompetition racingCarCompetition = new RacingCarCompetition();
+        racingCarCompetition.startRacing();
+    }
+}
