+++ src/main/java/racingcarcompetition/RandomGenerator.java
@@ -1,0 +1,6 @@
+package racingcarcompetition;
+
+@FunctionalInterface
+public interface RandomGenerator {
+    int generateRandom();
+}
