+++ src/main/java/racinggamewinner/Position.java
@@ -1,0 +1,17 @@
+package racinggamewinner;
+
+public class Position {
+    private int position;
+
+    public Position(int positionNumber) {
+        position = positionNumber;
+    }
+
+    public void updatePosition(int movement) {
+        position += movement;
+    }
+
+    public int currentPosition() {
+        return position;
+    }
+}
