+++ src/main/java/racinggamewinner/random/RacingCarRandomGenerator.java
@@ -1,0 +1,11 @@
+package racinggamewinner.random;
+
+import java.util.Random;
+
+public class RacingCarRandomGenerator implements RandomGenerator {
+    @Override
+    public int generateRandomNumber() {
+        Random random = new Random();
+        return random.nextInt(10);
+    }
+}
