+++ src/main/java/racinggamewinner/random/RandomGenerator.java
@@ -1,0 +1,6 @@
+package racinggamewinner.random;
+
+@FunctionalInterface
+public interface RandomGenerator {
+    int generateRandomNumber();
+}
