--- src/main/java/step3racingcarcompetition/Main.java
@@ -1,8 +1,0 @@
-package step3racingcarcompetition;
-
-public class Main {
-    public static void main(String[] args) {
-        RacingCarCompetition racingCarCompetition = new RacingCarCompetition();
-        racingCarCompetition.startRacing();
-    }
-}
