--- src/main/java/step3racingcarcompetition/RandomGenerator.java
@@ -1,6 +1,0 @@
-package step3racingcarcompetition;
-
-@FunctionalInterface
-public interface RandomGenerator {
-    int generateRandom();
-}
