--- src/main/java/step3racingcarcompetition/RandomNumber.java
@@ -1,17 +1,0 @@
-package step3racingcarcompetition;
-
-import java.util.Random;
-
-public class RandomNumber implements RandomGenerator {
-    private final int endRange;
-
-    public RandomNumber(int endRange) {
-        this.endRange = endRange;
-    }
-
-    @Override
-    public int generateRandom() {
-        Random random = new Random();
-        return random.nextInt(endRange);
-    }
-}
