--- src/main/java/step4racinggamewinner/Position.java
@@ -1,17 +1,0 @@
-package step4racinggamewinner;
-
-public class Position {
-    private int position;
-
-    public Position(int positionNumber) {
-        position = positionNumber;
-    }
-
-    public void updatePosition(int movement) {
-        position += movement;
-    }
-
-    public int currentPosition() {
-        return position;
-    }
-}
