--- src/main/java/step4racinggamewinner/random/RacingCarRandomGenerator.java
@@ -1,11 +1,0 @@
-package step4racinggamewinner.random;
-
-import java.util.Random;
-
-public class RacingCarRandomGenerator implements RandomGenerator {
-    @Override
-    public int generateRandomNumber() {
-        Random random = new Random();
-        return random.nextInt(10);
-    }
-}
