--- src/main/java/step4racinggamewinner/random/RandomGenerator.java
@@ -1,6 +1,0 @@
-package step4racinggamewinner.random;
-
-@FunctionalInterface
-public interface RandomGenerator {
-    int generateRandomNumber();
-}
