+++ src/test/java/racingcarcompetition/TestRandomNumberGenerator.java
@@ -1,0 +1,9 @@
+package racingcarcompetition;
+
+public class TestRandomNumberGenerator implements RandomGenerator {
+
+    @Override
+    public int generateRandom() {
+        return 4;
+    }
+}
