--- src/test/java/step2stringaddcalculator/StringAddCalculatorTest.java
@@ -1,50 +1,0 @@
-package step2stringaddcalculator;
-
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.NullAndEmptySource;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
-class StringAddCalculatorTest {
-
-    int result;
-
-    @ParameterizedTest
-    @NullAndEmptySource()
-    void splitAndSum_null_또는_빈문자(String input) {
-        result = StringAddCalculator.splitAndSum(input);
-        assertThat(result).isEqualTo(0);
-    }
-
-    @Test
-    void splitAndSum_숫자하나() {
-        result = StringAddCalculator.splitAndSum("1");
-        assertThat(result).isEqualTo(1);
-    }
-
-    @Test
-    void splitAndSum_쉼표구분자() {
-        result = StringAddCalculator.splitAndSum("1,2");
-        assertThat(result).isEqualTo(3);
-    }
-
-    @Test
-    void splitAndSum_쉼표_또는_콜론_구분자() {
-        result = StringAddCalculator.splitAndSum("1,2:3");
-        assertThat(result).isEqualTo(6);
-    }
-
-    @Test
-    void splitAndSum_custom_구분자() {
-        result = StringAddCalculator.splitAndSum("//;\n1;2;3");
-        assertThat(result).isEqualTo(6);
-    }
-
-    @Test
-    void splitAndSum_custom_negative() throws Exception {
-        assertThatThrownBy(() -> StringAddCalculator.splitAndSum("-1,2,3"))
-                .isInstanceOf(RuntimeException.class);
-    }
-}
