--- src/test/java/step3racingcarcompetition/TestRandomNumberGenerator.java
@@ -1,9 +1,0 @@
-package step3racingcarcompetition;
-
-public class TestRandomNumberGenerator implements RandomGenerator {
-
-    @Override
-    public int generateRandom() {
-        return 4;
-    }
-}
