--- src/test/java/step4racinggamewinner/AlwaysMoveNumber.java
@@ -1,13 +1,0 @@
-package step4racinggamewinner;
-
-import step4racinggamewinner.random.RandomGenerator;
-
-public class AlwaysMoveNumber implements RandomGenerator {
-
-    public static final int TEST_MOVING_NUMBER = 4;
-
-    @Override
-    public int generateRandomNumber() {
-        return TEST_MOVING_NUMBER;
-    }
-}
